Some comments removed
